% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinysync-package.R
\docType{package}
\name{shinysync-package}
\alias{shinysync}
\alias{shinysync-package}
\title{shinysync: Collaborative Shiny Applications with Automerge}
\description{
Real-time collaborative widgets for Shiny applications, built on 'Automerge' 'CRDT'. Provides a 'CodeMirror' 6 code editor with WebSocket sync, collaborative textareas and kanban boards, general input synchronization across sessions, and a replay timeline for reviewing interaction history.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shikokuchuo/shinysync}
  \item \url{https://shikokuchuo.net/shinysync/}
  \item \url{http://shikokuchuo.net/shinysync/}
  \item Report bugs at \url{https://github.com/shikokuchuo/shinysync/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
